+++ src/resources/cpfieldinspect.js
@@ -1,6 +1,6 @@
 (function (window) {
 
-    if (!window.Craft || !window.jQuery) {
+    if (!window.Craft || !window.$) {
         return false;
     }
 
@@ -58,16 +58,16 @@
 
             // Add event handlers
             Garnish.$doc
-                .on('click', '.' + this.settings.settingsClassSelector + ' a', this.onCpFieldLinksClick.bind(this))
-                .on('click', '[data-cpfieldlinks-sourcebtn]', this.onCpFieldLinksClick.bind(this))
-                .on('click', '.matrix .btn.add, .matrix .btn[data-type]', this.onMatrixBlockAddButtonClick.bind(this))
-                .ajaxComplete(this.onAjaxComplete.bind(this));
+                .on('click', '[data-cpfieldlinks-sourcebtn]', $.proxy(this.onSourceEditBtnClick, this))
+                .on('click', '.matrix .btn.add, .matrix .btn[data-type]', $.proxy(this.onMatrixBlockAddButtonClick, this))
+                .ajaxComplete($.proxy(this.onAjaxComplete, this));
 
-            this.render();
+            Garnish.requestAnimationFrame($.proxy(this.addFieldLinks, this));
         },
+
         initElementEditor: function () {
             var now = new Date().getTime(),
-                doInitElementEditor = (function () {
+                doInitElementEditor = $.proxy(function () {
                     var timestamp = new Date().getTime(),
                         $elementEditor = $('.elementeditor:last'),
                         $hud = $elementEditor.length > 0 ? $elementEditor.closest('.hud') : false,
@@ -75,11 +75,11 @@
                     if (elementEditor && elementEditor.hud) {
                         this.elementEditors[elementEditor._namespace] = elementEditor;
                         elementEditor.hud.on('hide', $.proxy(this.destroyElementEditor, this, elementEditor));
-                        Garnish.requestAnimationFrame(this.addFieldLinks.bind(this));
+                        Garnish.requestAnimationFrame($.proxy(this.addFieldLinks, this));
                     } else if (timestamp - now < 2000) { // Poll for 2 secs
                         Garnish.requestAnimationFrame(doInitElementEditor);
                     }
-                }).bind(this);
+                }, this);
             doInitElementEditor();
         },
 
@@ -90,31 +90,21 @@
         },
 
         setPathAndRedirect: function () {
-
             var redirectTo = Craft.getLocalStorage(this.settings.redirectKey);
-
-            if (redirectTo)
-            {
-                var $actionInput = $('input[type="hidden"][name="action"]').filter(this.settings.actionInputKeys.join(',')),
-                    $redirectInput = $('input[type="hidden"][name="redirect"]');
-                if ($actionInput.length > 0 && $redirectInput.length > 0)
-                {
+            if (redirectTo) {
+                var $actionInput = $('input[type="hidden"][name="action"]').filter(this.settings.actionInputKeys.join(','));
+                var $redirectInput = $('input[type="hidden"][name="redirect"]');
+                if ($actionInput.length > 0 && $redirectInput.length > 0) {
                     $redirectInput.attr('value', redirectTo);
                 }
             }
             Craft.setLocalStorage(this.settings.redirectKey, null);
         },
 
-        render: function () {
-            $('[data-cpfieldlinks]').removeAttr('data-cpfieldlinks');
-            this.addFieldLinks();
-        },
-
         addFieldLinks: function () {
-
-            var self = this,
-                targets = [$(this.getFieldContextSelector())],
-                $target;
+            var self = this;
+            var targets = [$(this.getFieldContextSelector())];
+            var $target;
 
             if (this.elementEditors && Object.keys(this.elementEditors).length) {
                 for (var key in this.elementEditors) {
@@ -122,68 +112,71 @@
                 }
             }
 
+            if (!targets.length) {
+                return;
+            }
+
             for (var i = 0; i < targets.length; ++i) {
 
                 $target = targets[i];
-
                 if (!$target || !$target.length) {
                     continue;
                 }
 
-                // Add CpFieldLinks to regular fields
-                var fieldData = this.data.fields || {},
-                    $fields = $target.find('.field:not([data-cpfieldlinks])').not('.matrixblock .field'),
-                    $field,
-                    fieldHandle;
-                $fields.each(function () {
-                    $field = $(this);
-                    fieldHandle = self.getFieldHandleFromAttribute($field.attr('id'));
-                    if (fieldHandle && fieldData.hasOwnProperty(fieldHandle)) {
-                        $field.find('.heading:first label').after(self.templates.editFieldBtn(fieldData[fieldHandle]));
+                var _this = this;
+
+                var $copyFieldHandleButtons = $target.find('.field .heading [id$=-field-attribute].code:not([data-cpfieldlinks-inited])');
+                $copyFieldHandleButtons.each(function () {
+
+                    var $btn = $(this);
+                    $btn.attr('data-cpfieldlinks-inited', true);
+
+                    var field = $btn.parents('.field').get().slice(-1).pop();
+                    if (!field) {
+                        return;
                     }
-                    $field.attr('data-cpfieldlinks', true);
-                });
 
-                // Add CpFieldLinks to Commerce variant fields
-                $target.find('.variant-matrixblock:not([data-cpfieldlinks])').each(function () {
-                    $(this).attr('data-cpfieldlinks', true).find('.field').each(function () {
-                        $field = $(this);
-                        fieldHandle = self.getFieldHandleFromAttribute($field.attr('id'));
-                        if (fieldHandle && fieldData.hasOwnProperty(fieldHandle)) {
-                            $field.find('.heading:first label').after(self.templates.editFieldBtn(fieldData[fieldHandle]));
-                        }
-                    });
-                });
+                    var fieldId = _this.getFieldId(field);
+                    if (!fieldId) {
+                        return;
+                    }
 
-                // Add CpFieldLinks to Matrix blocks
-                var $matrixBlocks = $target.find('.matrixblock:not([data-cpfieldlinks])'),
-                    $block,
-                    blockId,
-                    blockFieldData;
-                $matrixBlocks.each(function () {
-                    $block = $(this);
-                    fieldHandle = self.getFieldHandleFromAttribute($block.closest('.field').attr('id'));
-                    if (!fieldHandle || !fieldData.hasOwnProperty(fieldHandle)) return;
-                    blockId = $block.data('id');
-                    $block.attr('data-cpfieldlinks', true).find('.field').each(function () {
-                        $field = $(this);
-                        blockFieldData = {
-                            id: fieldData[fieldHandle].id,
-                            handle: self.getFieldHandleFromAttribute($field.attr('id'))
-                        }
-                        $field.find('.heading:first label').after(self.templates.editFieldBtn(blockFieldData));
-                    });
+                    $btn
+                        .append('<span data-icon="settings" title="' + (_this.data.editFieldBtnLabel || 'Edit field settings') + '" />')
+                        .on('click', '[data-icon="settings"]', function (e) {
+                            e.preventDefault();
+                            e.stopPropagation();
+                            _this.redirectToFieldSettings(fieldId);
+                        });
+
+                }).on('mouseleave', function () {
+                    $(this).blur();
                 });
 
             }
-
         },
 
-        getFieldHandleFromAttribute: function (value) {
-            if (!value) return false;
-            value = value.split('-');
-            if (value.length < 3) return false;
-            return value[value.length-2];
+        doRedirect: function (href) {
+            Craft.setLocalStorage(this.settings.redirectKey, this.data.redirectUrl || null);
+            window.location.href = href;
+        },
+
+        redirectToFieldSettings: function (fieldId) {
+            var href = Craft.CpFieldInspectPlugin.data.baseEditFieldUrl + '/' + fieldId;
+            this.doRedirect(href);
+        },
+
+        getFieldId: function (field) {
+            var id = $(field).attr('id');
+            if (!id) {
+                return null;
+            }
+            var segments = id.split('-');
+            if (segments.length < 3) {
+                return null;
+            }
+            var handle = segments[segments.length - 2];
+            return (this.data.fields || {})[handle] || null;
         },
 
         getFieldContextSelector: function () {
@@ -193,52 +186,36 @@
             return '#main';
         },
 
-        templates: {
-            editFieldBtn: function (attributes)
-            {
-                return  '<div class="cp-field-inspect cp-field-inspect-field-edit" aria-hidden="true">' +
-                    '<div class="' + Craft.CpFieldInspectPlugin.settings.settingsClassSelector + '">' +
-                    '<a href="' + Craft.CpFieldInspectPlugin.data.baseEditFieldUrl + '/' + attributes.id + '" class="settings icon" role="button" aria-label="Edit field" tabindex="-1"></a>' +
-                    '</div>' +
-                    '<div class="' + Craft.CpFieldInspectPlugin.settings.infoClassSelector + '">' + '<p><code>' + attributes.handle + '</code></p></div>' +
-                    '</div>';
-            }
-        },
-
         onLivePreviewEnter: function () {
             this.isLivePreview = true;
-            Garnish.requestAnimationFrame((function () {
-                this.addFieldLinks();
-            }).bind(this));
+            Garnish.requestAnimationFrame($.proxy(this.addFieldLinks, this));
         },
 
         onLivePreviewExit: function () {
             this.isLivePreview = false;
-            Garnish.requestAnimationFrame((function () {
-                this.addFieldLinks();
-            }).bind(this));
+            Garnish.requestAnimationFrame($.proxy(this.addFieldLinks, this));
         },
 
-        onCpFieldLinksClick: function (e) {
-            Craft.setLocalStorage(this.settings.redirectKey, this.data.redirectUrl || null);
+        onSourceEditBtnClick: function (e) {
+            e.preventDefault();
+            this.doRedirect(e.target.href);
         },
 
         onMatrixBlockAddButtonClick: function (e) {
-            Garnish.requestAnimationFrame((function () {
-                this.addFieldLinks();
-            }).bind(this));
+            Garnish.requestAnimationFrame($.proxy(this.addFieldLinks, this));
         },
 
         onAjaxComplete: function(e, status, requestData) {
-            if (requestData.url.indexOf('switch-entry-type') > -1) {
-                const $entryTypeSelect = $('#entryType');
-                if ($entryTypeSelect.length) {
-                    const typeId = $entryTypeSelect.val();
-                    $('[data-cpfieldlinks-sourcebtn][data-typeid]:not([data-typeid="' + typeId + '"]').hide();
-                    $('[data-cpfieldlinks-sourcebtn][data-typeid="' + typeId + '"]').show();
-                }
-                this.render();
+            if (requestData.url.indexOf('switch-entry-type') === -1) {
+                return;
             }
+            const $entryTypeSelect = $('#entryType');
+            if ($entryTypeSelect.length) {
+                const typeId = $entryTypeSelect.val();
+                $('[data-cpfieldlinks-sourcebtn][data-typeid]:not([data-typeid="' + typeId + '"]').hide();
+                $('[data-cpfieldlinks-sourcebtn][data-typeid="' + typeId + '"]').show();
+            }
+            this.addFieldLinks();
         }
     };